Multiply changes to code.
StringBuilder is a good choice when dealing with a large number of string concatenations, in this case, it's overkill. The number of concatenations is small and known in advance, so using string interpolation ($"{variable}") is simpler and more readable.

// Removed the unnecessary using directive.

// TicTacToe.cs line 78. Removed switch-case statement for code simplicity.
TicTacToe.cs:90 Fixed error message.

Added Board 5x5
Added Board property for cells in row for win. For example in level 5 need 4 cells in row for win.
Changed Check win logic.

Improved Board printing.